Refactor repo stuff into a single file; add unit tests
