+++ packages/runtime/package.json
@@ -40,6 +40,7 @@
     "README.md"
   ],
   "dependencies": {
-    "@openfn/logger": "workspace:^0.0.4"
+    "@openfn/logger": "workspace:^0.0.4",
+    "semver": "^7.3.8"
   }
 }
