+++ packages/runtime/src/index.ts
@@ -1,6 +1,4 @@
 import run from './runtime';
 export default run;
 
-import install from './repo/install';
-import { getModulePath } from './repo/util';
-export { install, getModulePath };
+export * from './modules/repo';
